Initial Commit
Game is in a playable State. The first level is nearly complete and I am mostly happy with CJJ's Movement. There are some animations which need to be made and a few non-interactable sprites (i.e. the checkpoints) which need to be drawn. The second level exists, but is not designed at all.